+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Bug修复
 * 【http】          修复HttpRequest.setFollowRedirects无效问题（issue#ILIKG@Gitee）
 * 【core】         修复CharUtil.isEmoji问题
+* 【http】          修复HttpResponse.writeBody同步模式下写出失败问题
 
 -------------------------------------------------------------------------------------------------------------